+++ src/view/tasks_view.ts
@@ -3,14 +3,14 @@ import * as vscode from 'vscode';
 import * as taskModel from '../model/tasks_model';
 
 function taskDescription(task: taskModel.Task): string {
-    return task.state===taskModel.TaskState.R?task.runing_time +'/'+task.limit_time:task.reason;
+    return task.state === taskModel.TaskState.R ? task.runing_time + '/' + task.limit_time : task.reason;
 }
 
 function taskTooltip(task: taskModel.Task): vscode.MarkdownString {
     return new vscode.MarkdownString(`
-| id | name | user | status | gres |
-|:--:|:--:|:--:|:--:|:--:|
-| ${task.jobid} | ${task.name} | ${task.user} | ${task.state} | ${task.gres} |
+| id | user | status | gres |
+|:--:|:--:|:--:|:--:|
+| ${task.jobid} | ${task.user} | ${task.state} | ${task.gres} |
     `);
 }
 
@@ -22,6 +22,8 @@ const iconMap: Map<taskModel.TaskState, vscode.ThemeIcon> = new Map([
     // [taskModel.TaskState.PD, new vscode.ThemeIcon('pass', new vscode.ThemeColor('gitDecoration.deletedResourceForeground'))], //勾
 ]);
 
+const finishedIcon = new vscode.ThemeIcon('pass', new vscode.ThemeColor('gitDecoration.addedResourceForeground'));
+
 export class TaskViewItem extends vscode.TreeItem {
     constructor(public readonly task: taskModel.Task) {
         super(task.name, vscode.TreeItemCollapsibleState.Collapsed);
@@ -29,6 +31,22 @@ export class TaskViewItem extends vscode.TreeItem {
         this.tooltip = taskTooltip(task);
         this.iconPath = iconMap.get(task.state);
         this.contextValue = 'taskItem';
+    }
+}
+
+export class FinishedTaskViewItem extends vscode.TreeItem {
+    constructor(public readonly task: taskModel.Task) {
+        super(task.name, vscode.TreeItemCollapsibleState.Collapsed);
+        this.iconPath = finishedIcon;
+        this.contextValue = 'finishedTaskItem';
+    }
+}
+
+export class listItem extends vscode.TreeItem {
+    constructor(public readonly title: string, public readonly children: vscode.TreeItem[], contextValue?: string) {
+        super(title, vscode.TreeItemCollapsibleState.Collapsed);
+        this.description = `${children.length} tasks`;
+        this.contextValue = contextValue;
     }
 }
 
@@ -59,7 +77,18 @@ function getTaskInfoItems(task: taskModel.Task): vscode.TreeItem[] {
     ];
 }
 
-export class TaskViewDataProvider implements vscode.TreeDataProvider<TaskViewItem|InfoItem> {
+function getListItemsOfGroupedTask(tasks: taskModel.Task[]): listItem[] {
+    let running = tasks.filter(v => !v.finished);
+    let finished = tasks.filter(v => v.finished);
+    console.log(running)
+    console.log(finished)
+    return [
+        new listItem('running', running.map((value) => { return new TaskViewItem(value); })),
+        new listItem('finished', finished.map((value) => { return new TaskViewItem(value); })),
+    ];
+}
+
+export class TaskViewDataProvider implements vscode.TreeDataProvider<TaskViewItem | InfoItem> {
     private static _instance: TaskViewDataProvider | null = null;
     private constructor() { }
 
@@ -77,16 +106,18 @@ export class TaskViewDataProvider implements vscode.TreeDataProvider<TaskViewIte
         this._onDidChangeTreeData.fire();
     }
 
-    getTreeItem(element: TaskViewItem|InfoItem): vscode.TreeItem {
+    getTreeItem(element: TaskViewItem | InfoItem): vscode.TreeItem {
         return element;
     }
 
-    getChildren(element?: TaskViewItem|InfoItem): Thenable<TaskViewItem[]|InfoItem[]> {
+    getChildren(element?: TaskViewItem | InfoItem): Thenable<TaskViewItem[] | InfoItem[]> {
         if (!element) {
-            return Promise.resolve(taskModel.taskManager.getTask().map((value) => { return new TaskViewItem(value); }));
+            return Promise.resolve(getListItemsOfGroupedTask(taskModel.taskManager.getTask()));
         }
         if (element instanceof TaskViewItem) {
             return Promise.resolve(getTaskInfoItems(element.task));
+        } else if (element instanceof listItem) {
+            return Promise.resolve(element.children);
         }
         return Promise.resolve([]);
     }